fix: Resolve circular imports and reconcile with Phase 2.1
- Move CodeChunk to separate types.py module to avoid circular imports
- Create plugin_base.py as bridge between Phase 1.2 plugin system and Phase 2.1 config system
- Update all imports to use correct module locations
- Keep both PythonConfig (Phase 2.1) and PythonPlugin (Phase 1.2) for compatibility
