+++ chunker/__init__.py
@@ -31,12 +31,14 @@ from .parser import (
     get_parser, list_languages, get_language_info, 
     return_parser, clear_cache
 )
-from .chunker import chunk_file, CodeChunk
+from .types import CodeChunk
+from .chunker import chunk_file
 from .factory import ParserConfig
 from .exceptions import (
     ChunkerError, LanguageNotFoundError, 
     ParserError, LibraryNotFoundError
 )
 from .plugin_manager import PluginManager, get_plugin_manager
-from .languages.base import LanguagePlugin, PluginConfig
+from .languages.base import PluginConfig
+from .languages.plugin_base import LanguagePlugin
 from .config import ChunkerConfig
