+++ chunker/chunker.py
@@ -1,22 +1,10 @@
 from __future__ import annotations
 from pathlib import Path
-from dataclasses import dataclass
 from tree_sitter import Node
 
 from .parser import get_parser
 from .languages import language_config_registry
-
-@dataclass
-class CodeChunk:
-    language: str
-    file_path: str
-    node_type: str
-    start_line: int
-    end_line: int
-    byte_start: int
-    byte_end: int
-    parent_context: str
-    content: str
+from .types import CodeChunk
 
 def _walk(node: Node, source: bytes, language: str, parent_ctx: str | None = None) -> list[CodeChunk]:
     """Walk the AST and extract chunks based on language configuration."""
