+++ chunker/types.py
@@ -1,0 +1,16 @@
+"""Common data types for the chunker."""
+
+from dataclasses import dataclass
+
+
+@dataclass
+class CodeChunk:
+    language: str
+    file_path: str
+    node_type: str
+    start_line: int
+    end_line: int
+    byte_start: int
+    byte_end: int
+    parent_context: str
+    content: str
